start rewriting flight-service

+++ flight-service/src/main/java/com/flightService/repository/TicketRepository.java
@@ -8,10 +8,10 @@ import java.util.List;
 public interface TicketRepository extends GenericJPARepository<Ticket> {
 
     @Query("from Ticket t LEFT JOIN FETCH t.flight f where f.id= ?1")
-    List<Ticket> findTicketsByFlightId(Integer flightId);
+    List<Ticket> findTicketsByFlightId(Long flightId);
 
     @Query("from Ticket t LEFT JOIN FETCH t.passenger p where p.id= ?1")
-    List<Ticket> findTicketsByPassengerId(Integer passengerId);
+    List<Ticket> findTicketsByPassengerId(Long passengerId);
 
     List<Ticket> findByActiveTrue();
